+++ core/src/cpp/include/machine/gnarly.hpp
