+++ core/src/cpp/include/misc/scalar.hpp
