+++ core/src/cpp/mc64k.make
@@ -1,6 +1,6 @@
 # Common include for building the interpreter
 
-OBJ = obj/$(ARCH)/main.o obj/$(ARCH)/standard_test_host.o obj/$(ARCH)/host.o obj/$(ARCH)/interpreter.o obj/$(ARCH)/interpreter_ea.o obj/$(ARCH)/interpreter_run.o obj/$(ARCH)/loader.o
+OBJ = obj/$(ARCH)/main.o obj/$(ARCH)/standard_test_host.o obj/$(ARCH)/host.o obj/$(ARCH)/interpreter.o obj/$(ARCH)/interpreter_ea.o obj/$(ARCH)/interpreter_run.o obj/$(ARCH)/loader.o obj/$(ARCH)/misc.o
 
 $(BIN): $(OBJ) Makefile.$(MEXT)
 	$(CXX) $(CXXFLAGS) $(OBJ) -o $(BIN)
